chore(tanstack-react-start): TypeScript 6.0 compatibility
