+++ packages/tanstack-react-start/src/client/index.ts
@@ -1,3 +1,2 @@
 export * from './ClerkProvider';
 export { SignIn, SignUp, OrganizationProfile, OrganizationList, UserProfile } from './uiComponents';
-export { UNSAFE_PortalProvider } from '@clerk/react';
