+++ packages/tanstack-react-start/src/index.ts
@@ -1,4 +1,5 @@
-export * from './client/index';
+export * from '@clerk/react';
+export { ClerkProvider, SignIn, SignUp, OrganizationProfile, OrganizationList, UserProfile } from './client/index';
 export { getToken } from '@clerk/shared/getToken';
 
 // Override Clerk React error thrower to show that errors come from @clerk/tanstack-react-start
